transform/pyproject.toml
[project]
name = "transform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.9.0",
    "pyspark>=4.1.1",
]
